+++ src/validation/parseFile.spec.ts
@@ -57,6 +57,16 @@ describe('parseFile', () => {
       fileName: 'notAPermissionArray.json',
       error: 'permissions: value is not an array',
     },
+    {
+      title: 'should fail on permission not being an array',
+      fileName: 'notAPermissionArray.json',
+      error: 'permissions: value is not an array',
+    },
+    {
+      title: 'should fail on permission not being an array',
+      fileName: 'notAPermissionArray.json',
+      error: 'permissions: value is not an array',
+    },
   ])('$title', async ({fileName, error}) => {
     const file = join(pathToStub, 'jsondocuments', fileName);
     await expect(parse(file)).rejects.toThrow(