+++ src/validation/parsePermissions.ts
@@ -73,7 +73,6 @@ const processPermissionLevel = (
   documentBuilder: DocumentBuilder,
   documentPath: PathLike
 ) => {
-  // TODO: require at least some permissions
   const permissionSetBuilders = permission.permissionSets!.map(
     (permissionSet) => {
       const caseInsensitivePermissions = new CaseInsensitiveDocument(
@@ -97,7 +96,6 @@ const processPermissionLevel = (
     }
   );
 
-  // TODO: prevent empty permission set
   documentBuilder.withPermissionLevel(permission.name!, permissionSetBuilders);
 };